Added act_fn optionality to td3 and sac
